Create values manager + refactor elements
https://coveord.atlassian.net/browse/JSUI-2373
